+++ src/HoudiniBifrost.h
@@ -22,28 +22,12 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#ifndef GEO_BIFROSTIOTRANSLATOR_H
-#define GEO_BIFROSTIOTRANSLATOR_H
+#ifndef HOUDINIBIFROST_H
+#define HOUDINIBIFROST_H
 
-#include "GEO/GEO_IOTranslator.h"
+namespace Bifrost{namespace API{class Component;}};
+class GU_Detail;
 
-class GEO_BifrostIOTranslator : public GEO_IOTranslator
-{
-public:
-	GEO_BifrostIOTranslator();
-	GEO_BifrostIOTranslator(const GEO_BifrostIOTranslator& other);
-	virtual ~GEO_BifrostIOTranslator();
+bool convertBifrostPointCloud(const Bifrost::API::Component& component, GU_Detail& geo);
 
-	virtual GEO_IOTranslator* duplicate() const;
-	virtual const char*	formatName() const;
-	virtual int	checkExtension(const char *name);
-
-	virtual int	checkMagicNumber(unsigned magic);
-
-	virtual GA_Detail::IOStatus fileLoad(GEO_Detail *gdp, UT_IStream &is, bool ate_magic);
-	virtual GA_Detail::IOStatus fileSave(const GEO_Detail *gdp, std::ostream &os);
-
-	static void install();
-};
-
-#endif//GEO_BIFROSTIOTRANSLATOR_H
+#endif//HOUDINIBIFROST_H
